requirements.txt
# Telegram Bot
python-telegram-bot==21.0

# Embeddings & Vector DB
sentence-transformers==2.3.1
chromadb==0.4.22

# Local LLM (Ollama)
ollama==0.2.0

# Vision Models
transformers==4.36.2
torch==2.1.2
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3